fix(webhook): address review feedback on rate-limit bounds and multipart gap
- Fix ConnectorProperties.Webhook.maxRequestBodyBytes type mismatch: it was
  long, but the controller field it documents is int (readNBytes takes an
  int). A long-typed value from metadata could bind a value the runtime
  can't accept.
- Bound the per-executable rate-limiter map with a Guava Cache
  (maximumSize + expireAfterAccess) instead of a plain ConcurrentHashMap.
  ExecutableId is derived from the deduplication id, which includes the
  process-definition key, so redeployments mint new ids and the previous
  map implementation would accumulate one stale RateLimiter per
  redeployment for the controller's lifetime.
- Validate camunda.connector.webhook.rate-limit.permits-per-second at
  startup (@PostConstruct) when the rate limit is enabled, so a
  misconfigured non-positive value fails fast at boot instead of throwing
  on the first request to a registered webhook.
- Pin spring.servlet.multipart.max-file-size/max-request-size explicitly
  in the three shipped application.properties. Spring's DispatcherServlet
  parses multipart/form-data before the handler runs (and therefore before
  this fix's path/rate/size checks), bounded only by these limits, which
  previously relied on Spring Boot's implicit defaults rather than being
  pinned.
- Scope the unknown-path test name/comments to non-multipart requests,
  matching what the test actually proves.

Addresses github-copilot review feedback on camunda/connectors#8987 /
security-testing-findings#268.

+++ connector-runtime/spring-boot-starter-camunda-connectors/src/main/java/io/camunda/connector/runtime/ConnectorProperties.java
@@ -38,14 +38,15 @@ public record ConnectorProperties(
    *
    * @param maxRequestBodyBytes Maximum size, in bytes, of an inbound webhook request body read into
    *     memory before authentication runs. Requests whose body exceeds this limit are rejected with
-   *     413 Payload Too Large. Default is 10 MB (10485760).
+   *     413 Payload Too Large. Default is 10 MB (10485760). {@code int}-typed to match the
+   *     underlying {@code InputStream.readNBytes(int)} call it configures.
    * @param rateLimit Per-registered-webhook inbound rate limit, enforced before the request body is
    *     read.
    */
   public record Webhook(
       boolean enabled,
       boolean appendPhysicalTenantAndTenantToPath,
-      long maxRequestBodyBytes,
+      int maxRequestBodyBytes,
       RateLimit rateLimit) {}
 
   /**
